Add files via upload
As part of harnessing my skills in data analysis and getting comfortable in machine learning, I decided to work on this simple regression task that aims to predict car fuel consumption. In this project, I explored my data, identified various insights about the features with representation as well as delving in data modelling to make informed decisions.